Last verision

+++ API/src/moduels/message/message.controller.js
@@ -10,7 +10,7 @@ const addMessage = async (req, res, next) => {
         receiverId,
     })
     // const message = await Message.insertMany(req.body)
-    res.status(201).json({ message: "Message added sucessfully", sucess: true })
+    res.status(201).json({ message: "Message added sucessfully", success: true })
 }
 
 //get all message for this user
@@ -23,9 +23,18 @@ const getAllMessage = async (req, res, next) => {
 // delete message
 const deleteMessage = async (req, res, next) => {
     const messageId = req.params.id
-    const message = await Message.findByIdAndDelete(messageId)
-    res.status(200).json({ message: "Message deleted sucessfully", sucess: true })
+    const messageExist = await Message.findById(messageId)
+    if (!messageExist) {
+        return res.status(404).json({ message: "Message not found", success: false })
     }
+    // check if userId = reciverId login
+    if (messageExist.receiverId.toString() !== req.user.id) {
+        return res.status(403).json({ message: "You are not authorized to delete this message ", success: false })
+    }
+    // delete message
+    const message = await Message.findByIdAndDelete(messageId)
+    res.status(200).json({ message: "Message deleted successfully", success: true })
+}
